# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Configuration System API
  version: 2.0.0
  description: 保留 v1 发布图片链，并提供消费根 contracts SC01 草案 fixture 的 v2 配置 API。
servers:
  - url: /
paths:
  /health:
    get:
      operationId: getHealth
      summary: 健康检查
      responses:
        "200":
          description: 服务和已发布数据可用
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [status, catalogVersion, publicationVersion]
                properties:
                  status:
                    const: ok
                  catalogVersion:
                    type: string
                  publicationVersion:
                    type: string
  /api/v1/catalog:
    get:
      operationId: getCatalog
      summary: 获取当前选配目录
      responses:
        "200":
          description: 当前目录
          content:
            application/json:
              schema:
                $ref: ./schemas/catalog.schema.json
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/renders/resolve:
    post:
      operationId: resolveRender
      summary: 将合法配置和视角解析为渲染图片 URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveRenderRequest"
      responses:
        "200":
          description: 图片解析成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolveRenderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/VersionConflict"
  /api/v2/catalog:
    get:
      operationId: getCatalogV2
      summary: 获取 SC01 v2 草案目录
      responses:
        "200":
          description: 根 contracts 中的 SC01 v2 catalog fixture
          content:
            application/json:
              schema:
                $ref: ./schemas/catalog.v2.schema.json
  /api/v2/configurations:
    post:
      operationId: createConfigurationV2
      summary: 创建或幂等复用 SC01 v2 配置
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigurationWriteV2"
      responses:
        "201":
          description: 配置已创建，revision 从 1 开始
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoredConfigurationV2"
        "200":
          description: 相同幂等键或相同稳定配置的重放结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoredConfigurationV2"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/PriceBlocked"
    get:
      operationId: listOrGetConfigurationsV2
      summary: 列出配置，或按 query 中的 configurationId 获取配置
      parameters:
        - name: configurationId
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ConfigurationIdV2"
      responses:
        "200":
          description: 配置或配置列表
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateConfigurationV2
      summary: 按请求体 configurationId 和 revision 更新配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigurationUpdateV2"
      responses:
        "200":
          description: 更新成功，revision 加一
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoredConfigurationV2"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/RevisionConflict"
        "422":
          $ref: "#/components/responses/PriceBlocked"
  /api/v2/configurations/{configurationId}:
    parameters:
      - $ref: "#/components/parameters/ConfigurationIdV2"
    get:
      operationId: getConfigurationV2
      summary: 获取单个 SC01 v2 配置
      responses:
        "200":
          description: 已保存配置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoredConfigurationV2"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateConfigurationByIdV2
      summary: 使用乐观 revision 更新 SC01 v2 配置
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigurationUpdateByIdV2"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoredConfigurationV2"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/RevisionConflict"
        "422":
          $ref: "#/components/responses/PriceBlocked"
  /api/v2/renders/resolve:
    post:
      operationId: resolveRenderV2
      summary: 从完整选择解析稳定 configurationId 与 renderKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveRenderRequestV2"
      responses:
        "200":
          description: 渲染投影解析成功；SC01 草案不承诺已发布 imageUrl
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolveRenderResponseV2"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/VersionConflict"
  /assets/renders/{publicationVersion}/{vehicleId}/{configurationKey}/{renderViewId}.png:
    get:
      operationId: getRender
      summary: 获取已发布的 straight Alpha 透明 PNG
      description: 图片由 path-tracing 烘焙并按 manifest 的 alphaProcessing 策略归一化；ready render 会记录 coverage 反转、归一化和发光像素恢复结果。
      parameters:
        - $ref: "#/components/parameters/PublicationVersion"
        - $ref: "#/components/parameters/VehicleId"
        - $ref: "#/components/parameters/ConfigurationKey"
        - $ref: "#/components/parameters/RenderViewId"
      responses:
        "200":
          description: 已完成 P0-3 Alpha 处理的 sRGB RGBA PNG（straight Alpha）
          headers:
            Cache-Control:
              schema:
                type: string
              description: 不可变版本资源应使用长期缓存。
          content:
            image/png:
              schema:
                type: string
                contentEncoding: binary
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 同一键和同一规范配置返回原结果；同一键对应不同配置返回 409。
      schema:
        type: string
        minLength: 1
        maxLength: 128
    ConfigurationIdV2:
      name: configurationId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ConfigurationIdV2"
    PublicationVersion:
      name: publicationVersion
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/StableId"
    VehicleId:
      name: vehicleId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/StableId"
    ConfigurationKey:
      name: configurationKey
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ConfigurationKey"
    RenderViewId:
      name: renderViewId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/RenderViewId"
  schemas:
    ConfigurationIdV2:
      type: string
      pattern: ^cfg-[a-f0-9]{24}$
    SelectionsV2:
      type: object
      additionalProperties: false
      required:
        [exterior-body-cover, wheel-material, steering-wheel-skin]
      properties:
        exterior-body-cover:
          $ref: "#/components/schemas/StableId"
        wheel-material:
          $ref: "#/components/schemas/StableId"
        steering-wheel-skin:
          $ref: "#/components/schemas/StableId"
    CustomizationsV2:
      type: object
      additionalProperties:
        oneOf:
          - type: object
            additionalProperties: false
            required: [materialVariantId]
            properties:
              materialVariantId:
                $ref: "#/components/schemas/StableId"
          - type: object
            additionalProperties: false
            required: [colorHex, metallic, roughness, clearCoat, orangePeel, flakeIntensity]
            properties:
              colorHex:
                type: string
                pattern: ^#[0-9A-F]{6}$
              metallic:
                $ref: "#/components/schemas/UnitInterval"
              roughness:
                $ref: "#/components/schemas/UnitInterval"
              clearCoat:
                $ref: "#/components/schemas/UnitInterval"
              orangePeel:
                $ref: "#/components/schemas/UnitInterval"
              flakeIntensity:
                $ref: "#/components/schemas/UnitInterval"
    ConfigurationWriteV2:
      type: object
      additionalProperties: false
      required: [catalogVersion, vehicleId, selections]
      properties:
        catalogVersion:
          $ref: "#/components/schemas/StableId"
        vehicleId:
          $ref: "#/components/schemas/StableId"
        selections:
          $ref: "#/components/schemas/SelectionsV2"
        customizations:
          $ref: "#/components/schemas/CustomizationsV2"
        quoteRequested:
          type: boolean
          default: false
          description: SC01 draft 为 true 时返回 422，不生成伪报价。
    ConfigurationUpdateV2:
      type: object
      additionalProperties: false
      required:
        [configurationId, revision, catalogVersion, vehicleId, selections]
      properties:
        configurationId:
          $ref: "#/components/schemas/ConfigurationIdV2"
        revision:
          type: integer
          minimum: 1
        catalogVersion:
          $ref: "#/components/schemas/StableId"
        vehicleId:
          $ref: "#/components/schemas/StableId"
        selections:
          $ref: "#/components/schemas/SelectionsV2"
        customizations:
          $ref: "#/components/schemas/CustomizationsV2"
        quoteRequested:
          type: boolean
          default: false
    ConfigurationUpdateByIdV2:
      type: object
      additionalProperties: false
      required: [revision, catalogVersion, vehicleId, selections]
      properties:
        configurationId:
          $ref: "#/components/schemas/ConfigurationIdV2"
        revision:
          type: integer
          minimum: 1
        catalogVersion:
          $ref: "#/components/schemas/StableId"
        vehicleId:
          $ref: "#/components/schemas/StableId"
        selections:
          $ref: "#/components/schemas/SelectionsV2"
        customizations:
          $ref: "#/components/schemas/CustomizationsV2"
        quoteRequested:
          type: boolean
          default: false
    StoredConfigurationV2:
      type: object
      additionalProperties: false
      required:
        [schemaVersion, catalogVersion, vehicleId, configurationId, renderKey,
         selections, customizations, revision, priceResult, createdAt, updatedAt]
      properties:
        schemaVersion:
          const: 2.0.0
        catalogVersion:
          $ref: "#/components/schemas/StableId"
        vehicleId:
          $ref: "#/components/schemas/StableId"
        configurationId:
          $ref: "#/components/schemas/ConfigurationIdV2"
        renderKey:
          type: string
          pattern: ^[a-z0-9-]+__[a-z0-9-]+__render-[a-f0-9]{24}$
        selections:
          $ref: "#/components/schemas/SelectionsV2"
        customizations:
          $ref: "#/components/schemas/CustomizationsV2"
        revision:
          type: integer
          minimum: 1
        priceResult:
          $ref: ./schemas/price-result.v2.schema.json
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ResolveRenderRequestV2:
      type: object
      additionalProperties: false
      required: [catalogVersion, vehicleId, selections]
      properties:
        catalogVersion:
          $ref: "#/components/schemas/StableId"
        vehicleId:
          $ref: "#/components/schemas/StableId"
        selections:
          $ref: "#/components/schemas/SelectionsV2"
        customizations:
          $ref: "#/components/schemas/CustomizationsV2"
        renderViewId:
          $ref: "#/components/schemas/StableId"
    ResolveRenderResponseV2:
      type: object
      additionalProperties: false
      required:
        [schemaVersion, catalogVersion, vehicleId, configurationId, renderKey]
      properties:
        schemaVersion:
          const: 2.0.0
        catalogVersion:
          $ref: "#/components/schemas/StableId"
        vehicleId:
          $ref: "#/components/schemas/StableId"
        configurationId:
          $ref: "#/components/schemas/ConfigurationIdV2"
        renderKey:
          type: string
          pattern: ^[a-z0-9-]+__[a-z0-9-]+__render-[a-f0-9]{24}$
        renderViewId:
          $ref: "#/components/schemas/StableId"
    StableId:
      type: string
      pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
    UnitInterval:
      type: number
      minimum: 0
      maximum: 1
    ConfigurationKey:
      type: string
      pattern: ^paint-[a-z0-9-]+__wheel-[a-z0-9-]+__interior-[a-z0-9-]+__frame-[a-z0-9-]+$
    RenderViewId:
      type: string
      enum: [front, front-left, side, rear-right]
    Selections:
      type: object
      additionalProperties: false
      required: [paint, wheel, interior, frame]
      properties:
        paint:
          $ref: "#/components/schemas/StableId"
        wheel:
          $ref: "#/components/schemas/StableId"
        interior:
          $ref: "#/components/schemas/StableId"
        frame:
          $ref: "#/components/schemas/StableId"
    ResolveRenderRequest:
      type: object
      additionalProperties: false
      required:
        [catalogVersion, publicationVersion, vehicleId, selections, renderViewId]
      properties:
        catalogVersion:
          $ref: "#/components/schemas/StableId"
        publicationVersion:
          $ref: "#/components/schemas/StableId"
        vehicleId:
          $ref: "#/components/schemas/StableId"
        selections:
          $ref: "#/components/schemas/Selections"
        renderViewId:
          $ref: "#/components/schemas/RenderViewId"
    ResolveRenderResponse:
      type: object
      additionalProperties: false
      required:
        [configurationKey, renderViewId, imageUrl]
      properties:
        configurationKey:
          $ref: "#/components/schemas/ConfigurationKey"
        renderViewId:
          $ref: "#/components/schemas/RenderViewId"
        imageUrl:
          type: string
          pattern: ^/assets/renders/
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: 请求 JSON、ID、配置或视角非法
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: manifest 中不存在对应图片，或对应 render 尚未达到 ready 状态
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    VersionConflict:
      description: 客户端版本与当前发布版本冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: catalog 版本或幂等键冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RevisionConflict:
      description: 客户端 revision 不是当前 revision
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    PriceBlocked:
      description: SC01 draft 金额未确认，quoteAllowed=false，禁止报价
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: catalog 或满足 P0-3 renderer、alphaProcessing 与 ready render 元数据要求的 manifest 尚未就绪
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"